+++ internal/updater/updater.go
@@ -1,0 +1,165 @@
+package updater
+
+import (
+	"errors"
+	"fmt"
+	"io"
+	"path"
+
+	"github.com/Gelio/go-global-update/internal/gobinaries"
+	"github.com/Gelio/go-global-update/internal/gocli"
+	"go.uber.org/zap"
+)
+
+type Options struct {
+	Debug   bool
+	Verbose bool
+	DryRun  bool
+	// List of binary names to update.
+	// If empty, will update all binaries in GOBIN
+	BinariesToUpdate []string
+}
+
+// TODO: add tests for the updater
+
+// UpdateBinaries updates binaries in GOBIN
+//
+// If binariesToUpdate is empty, the command will attempt to update all
+// found binaries in GOBIN.
+func UpdateBinaries(
+	logger *zap.Logger,
+	options Options,
+	out io.Writer,
+	cmdRunner gocli.GoCmdRunner,
+	lister gobinaries.DirectoryLister,
+	fs FilesystemUtils,
+) error {
+	goCLI := gocli.New(cmdRunner)
+	gobin, err := getExecutableBinariesPath(&goCLI)
+	if err != nil {
+		return fmt.Errorf("could not determine GOBIN path: %w", err)
+	}
+
+	logger.Debug("found GOBIN path", zap.String("GOBIN", gobin))
+
+	if err := fs.Chdir(gobin); err != nil {
+		return fmt.Errorf("could not change directory to GOBIN (%s): %w", gobin, err)
+	}
+
+	introspecter := gobinaries.NewIntrospecter(cmdRunner, gobin, logger)
+	binaryNames, err := resolveBinaryNames(options.BinariesToUpdate, lister, gobin)
+	if err != nil {
+		return err
+	}
+
+	introspectionResults := gobinaries.IntrospectBinaries(&introspecter, binaryNames)
+	printBinariesSummary(introspectionResults, out, options.Verbose)
+
+	if !options.DryRun {
+		fmt.Fprintln(out)
+		return updateBinaries(introspectionResults, &goCLI, out, options.Verbose)
+	}
+
+	return nil
+}
+
+func resolveBinaryNames(binariesToUpdate []string, lister gobinaries.DirectoryLister, gobin string) ([]string, error) {
+	if len(binariesToUpdate) > 0 {
+		return binariesToUpdate, nil
+	}
+
+	binaryNames, err := lister.ListDirectoryEntries(gobin)
+	if err != nil {
+		err = fmt.Errorf("could not list GOBIN (%s) entries: %w", gobin, err)
+	}
+
+	return binaryNames, err
+}
+
+func printBinariesSummary(
+	introspectionResults []gobinaries.IntrospectionResult,
+	out io.Writer,
+	verbose bool,
+) {
+	for _, result := range introspectionResults {
+		if result.Error != nil {
+			fmt.Fprintln(out, result.Error)
+			continue
+		}
+
+		binary := result.Binary
+		var latestVersionInfo string
+		if binary.UpgradePossible() {
+			latestVersionInfo = fmt.Sprintf("can upgrade to %s", binary.LatestVersion)
+		} else {
+			latestVersionInfo = "up-to-date"
+		}
+
+		name := binary.Name
+		if verbose {
+			name = binary.PathURL
+		}
+
+		fmt.Fprintf(out, "%s (version: %s, %s)\n", name, binary.Version, latestVersionInfo)
+	}
+}
+
+func updateBinaries(
+	introspectionResults []gobinaries.IntrospectionResult,
+	goCLI *gocli.GoCLI,
+	out io.Writer,
+	verbose bool,
+) error {
+	var upgradeErrors []error
+
+	for _, result := range introspectionResults {
+		if result.Error != nil || !result.Binary.UpgradePossible() {
+			continue
+		}
+
+		binary := result.Binary
+
+		fmt.Fprintf(out, "Upgrading %s to %s ... ", binary.Name, binary.LatestVersion)
+		upgradeOutput, err := goCLI.UpgradePackage(binary.PathURL)
+		if err != nil {
+			upgradeErrors = append(upgradeErrors, err)
+			fmt.Fprintln(out, "❌")
+			fmt.Fprintln(out, "\tCould not upgrade package")
+		} else {
+			fmt.Fprintln(out, "✅")
+		}
+
+		if len(upgradeOutput) > 0 && (verbose || err != nil) {
+			fmt.Fprintln(out, upgradeOutput)
+		}
+		fmt.Fprintln(out)
+	}
+
+	if len(upgradeErrors) > 0 {
+		return fmt.Errorf("could not upgrade %d packages", len(upgradeErrors))
+	}
+
+	return nil
+}
+
+func getExecutableBinariesPath(cli *gocli.GoCLI) (string, error) {
+	gobin, err := cli.GetEnvVar("GOBIN")
+	if err != nil {
+		return "", nil
+	}
+	if len(gobin) > 0 {
+		return gobin, nil
+	}
+
+	gopath, err := cli.GetEnvVar("GOPATH")
+	if err != nil {
+		return "", nil
+	}
+	if len(gopath) == 0 {
+		return "", errors.New("GOPATH and GOPATH are not defined in 'go env' command")
+	}
+
+	gobin = path.Join(gopath, "bin")
+
+	return gobin, nil
+}