+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 Initial release.
 
 - Complete the basic functionality of upgrading globally installed executables.
+- Add `--dry-run`, `--verbose`, `--debug` flags.
+- Support upgrading only specified binaries by accepting arguments.